SLING-10931 - move util back to test - it's not needed for repoinit
