+++ src/test/java/org/apache/sling/feature/cpconverter/accesscontrol/EnforcePrincipalBasedTest.java
@@ -43,7 +43,7 @@ import java.util.Dictionary;
 import java.util.HashMap;
 import java.util.List;
 
-import static org.apache.sling.feature.cpconverter.repoinit.Util.normalize;
+import static org.apache.sling.feature.cpconverter.Util.normalize;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
